Rename test tot tests, update chiptool main
